requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
sentence-transformers
numpy
openai
python-dotenv
